pyproject.toml
[project]
name = "email-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "geonamescache>=2.0.0",
    "google-genai>=1.2.0",
    "markdownify>=0.14.1",
    "matplotlib>=3.10.0",
    "notebook>=7.3.2",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "scipy>=1.15.1",
]
